move entity warning and tap/hold handling to generic row element
