+++ src/panels/lovelace/entity-rows/hui-input-select-entity-row.ts
@@ -13,15 +13,12 @@ import "@polymer/paper-item/paper-item";
 import "@polymer/paper-listbox/paper-listbox";
 
 import "../../../components/entity/state-badge";
-import "../components/hui-warning";
 import "../components/hui-generic-entity-row";
 
 import computeStateName from "../../../common/entity/compute_state_name";
 import { HomeAssistant } from "../../../types";
 import { EntityRow, EntityConfig } from "./types";
 import { setOption } from "../../../data/input-select";
-import { longPress } from "../common/directives/long-press-directive";
-import { handleClick } from "../common/handle-click";
 
 @customElement("hui-input-select-entity-row")
 class HuiInputSelectEntityRow extends LitElement implements EntityRow {
@@ -44,25 +41,10 @@ class HuiInputSelectEntityRow extends LitElement implements EntityRow {
 
     const stateObj = this.hass.states[this._config.entity];
 
-    if (!stateObj) {
-      return html`
-        <hui-warning
-          >${this.hass.localize(
-            "ui.panel.lovelace.warning.entity_not_found",
-            "entity",
-            this._config.entity
-          )}</hui-warning
-        >
-      `;
-    }
-
     return html`
     <hui-generic-entity-row
         .hass="${this.hass}"
         .config="${this._config}"
-        @ha-click="${this._handleTap}"
-        @ha-hold="${this._handleHold}"
-        .longPress="${longPress()}"
       ></hui-generic-entity-row>
       <paper-dropdown-menu
         selected-item-label="${stateObj.state}"
@@ -111,14 +93,6 @@ class HuiInputSelectEntityRow extends LitElement implements EntityRow {
     }
 
     setOption(this.hass!, stateObj.entity_id, ev.target.selectedItem.innerText);
-  }
-
-  private _handleTap() {
-    handleClick(this, this.hass!, this._config!, false);
-  }
-
-  private _handleHold() {
-    handleClick(this, this.hass!, this._config!, true);
   }
 }
