+++ src/panels/lovelace/entity-rows/hui-lock-entity-row.ts
@@ -9,12 +9,9 @@ import {
 } from "lit-element";
 
 import "../components/hui-generic-entity-row";
-import "../components/hui-warning";
 
 import { HomeAssistant } from "../../../types";
 import { EntityRow, EntityConfig } from "./types";
-import { longPress } from "../common/directives/long-press-directive";
-import { handleClick } from "../common/handle-click";
 
 @customElement("hui-lock-entity-row")
 class HuiLockEntityRow extends LitElement implements EntityRow {
@@ -36,26 +33,8 @@ class HuiLockEntityRow extends LitElement implements EntityRow {
 
     const stateObj = this.hass.states[this._config.entity];
 
-    if (!stateObj) {
-      return html`
-        <hui-warning
-          >${this.hass.localize(
-            "ui.panel.lovelace.warning.entity_not_found",
-            "entity",
-            this._config.entity
-          )}</hui-warning
-        >
-      `;
-    }
-
     return html`
-      <hui-generic-entity-row
-        .hass="${this.hass}"
-        .config="${this._config}"
-        @ha-click="${this._handleTap}"
-        @ha-hold="${this._handleHold}"
-        .longPress="${longPress()}"
-      >
+      <hui-generic-entity-row .hass="${this.hass}" .config="${this._config}">
         <mwc-button @click="${this._callService}">
           ${stateObj.state === "locked"
             ? this.hass!.localize("ui.card.lock.unlock")
@@ -81,14 +60,6 @@ class HuiLockEntityRow extends LitElement implements EntityRow {
       stateObj.state === "locked" ? "unlock" : "lock",
       { entity_id: stateObj.entity_id }
     );
-  }
-
-  private _handleTap() {
-    handleClick(this, this.hass!, this._config!, false);
-  }
-
-  private _handleHold() {
-    handleClick(this, this.hass!, this._config!, true);
   }
 }
