+++ src/panels/lovelace/entity-rows/hui-text-entity-row.ts
@@ -9,14 +9,11 @@ import {
 } from "lit-element";
 
 import "../components/hui-generic-entity-row";
-import "../components/hui-warning";
 
 import computeStateDisplay from "../../../common/entity/compute_state_display";
 
 import { HomeAssistant } from "../../../types";
 import { EntityRow, EntityConfig } from "./types";
-import { longPress } from "../common/directives/long-press-directive";
-import { handleClick } from "../common/handle-click";
 
 @customElement("hui-text-entity-row")
 class HuiTextEntityRow extends LitElement implements EntityRow {
@@ -38,26 +35,8 @@ class HuiTextEntityRow extends LitElement implements EntityRow {
 
     const stateObj = this.hass.states[this._config.entity];
 
-    if (!stateObj) {
-      return html`
-        <hui-warning
-          >${this.hass.localize(
-            "ui.panel.lovelace.warning.entity_not_found",
-            "entity",
-            this._config.entity
-          )}</hui-warning
-        >
-      `;
-    }
-
     return html`
-      <hui-generic-entity-row
-        .hass="${this.hass}"
-        .config="${this._config}"
-        @ha-click="${this._handleTap}"
-        @ha-hold="${this._handleHold}"
-        .longPress="${longPress()}"
-      >
+      <hui-generic-entity-row .hass="${this.hass}" .config="${this._config}">
         <div>
           ${computeStateDisplay(
             this.hass!.localize,
@@ -75,14 +54,6 @@ class HuiTextEntityRow extends LitElement implements EntityRow {
         text-align: right;
       }
     `;
-  }
-
-  private _handleTap() {
-    handleClick(this, this.hass!, this._config!, false);
-  }
-
-  private _handleHold() {
-    handleClick(this, this.hass!, this._config!, true);
   }
 }
