+++ src/panels/lovelace/entity-rows/hui-toggle-entity-row.ts
@@ -8,14 +8,11 @@ import {
 
 import "../components/hui-generic-entity-row";
 import "../../../components/entity/ha-entity-toggle";
-import "../components/hui-warning";
 
 import computeStateDisplay from "../../../common/entity/compute_state_display";
 
 import { HomeAssistant } from "../../../types";
 import { EntityRow, EntityConfig } from "./types";
-import { longPress } from "../common/directives/long-press-directive";
-import { handleClick } from "../common/handle-click";
 
 @customElement("hui-toggle-entity-row")
 class HuiToggleEntityRow extends LitElement implements EntityRow {
@@ -37,26 +34,8 @@ class HuiToggleEntityRow extends LitElement implements EntityRow {
 
     const stateObj = this.hass.states[this._config.entity];
 
-    if (!stateObj) {
-      return html`
-        <hui-warning
-          >${this.hass.localize(
-            "ui.panel.lovelace.warning.entity_not_found",
-            "entity",
-            this._config.entity
-          )}</hui-warning
-        >
-      `;
-    }
-
     return html`
-      <hui-generic-entity-row
-        .hass="${this.hass}"
-        .config="${this._config}"
-        @ha-click="${this._handleTap}"
-        @ha-hold="${this._handleHold}"
-        .longPress="${longPress()}"
-      >
+      <hui-generic-entity-row .hass="${this.hass}" .config="${this._config}">
         ${stateObj.state === "on" || stateObj.state === "off"
           ? html`
               <ha-entity-toggle
@@ -75,14 +54,6 @@ class HuiToggleEntityRow extends LitElement implements EntityRow {
             `}
       </hui-generic-entity-row>
     `;
-  }
-
-  private _handleTap() {
-    handleClick(this, this.hass!, this._config!, false);
-  }
-
-  private _handleHold() {
-    handleClick(this, this.hass!, this._config!, true);
   }
 }
